update: renamed 'get_range' -> 'get_list'
