+++ src/tobox/application/box/mod.rs
@@ -1,11 +1,4 @@
-pub mod get_by_id;
-pub mod get_by_ids;
-pub mod get_range;
-pub mod get_self;
+
+pub mod get_list;
 pub mod create;
-pub mod update;
-pub mod update_self;
-pub mod send_confirm_code;
-pub mod confirm;
-pub mod get_range;
-pub mod reset_password;
+
